# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NovaSave API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DetectResultPlatform = typeof DetectResultPlatform[keyof typeof DetectResultPlatform];


export const DetectResultPlatform = {
  tiktok: 'tiktok',
  youtube: 'youtube',
  pinterest: 'pinterest',
  unknown: 'unknown',
} as const;

export type DetectResultContentType = typeof DetectResultContentType[keyof typeof DetectResultContentType];


export const DetectResultContentType = {
  video: 'video',
  audio: 'audio',
  image: 'image',
  playlist: 'playlist',
  shorts: 'shorts',
  reels: 'reels',
  pin: 'pin',
  unknown: 'unknown',
} as const;

export interface DetectResult {
  platform: DetectResultPlatform;
  contentType: DetectResultContentType;
  valid: boolean;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  thumbnail?: string | null;
}

export type DownloadInputQuality = typeof DownloadInputQuality[keyof typeof DownloadInputQuality];


export const DownloadInputQuality = {
  auto: 'auto',
  hd: 'hd',
  sd: 'sd',
  low: 'low',
  audio_only: 'audio_only',
  '360p': '360p',
  '720p': '720p',
  '1080p': '1080p',
} as const;

export type DownloadInputFormat = typeof DownloadInputFormat[keyof typeof DownloadInputFormat];


export const DownloadInputFormat = {
  mp4: 'mp4',
  mp3: 'mp3',
  auto: 'auto',
} as const;

export interface DownloadInput {
  url: string;
  quality?: DownloadInputQuality;
  format?: DownloadInputFormat;
}

export interface MediaItem {
  url: string;
  quality: string;
  format: string;
  label: string;
  /** @nullable */
  fileSize?: number | null;
}

export type DownloadJobStatus = typeof DownloadJobStatus[keyof typeof DownloadJobStatus];


export const DownloadJobStatus = {
  pending: 'pending',
  processing: 'processing',
  ready: 'ready',
  failed: 'failed',
} as const;

export interface DownloadJob {
  jobId: string;
  status: DownloadJobStatus;
  url: string;
  platform: string;
  contentType: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  author?: string | null;
  /** @nullable */
  duration?: number | null;
  mediaItems?: MediaItem[];
  /** @nullable */
  error?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
  retryCount?: number;
  /** @nullable */
  apiUsed?: string | null;
}

export interface HistoryEntry {
  jobId: string;
  url: string;
  platform: string;
  contentType: string;
  status: string;
  /** @nullable */
  title: string | null;
  /** @nullable */
  thumbnail: string | null;
  createdAt: string;
}

export type PublicStatsPlatformBreakdown = {
  tiktok?: number;
  youtube?: number;
  pinterest?: number;
};

export interface PublicStats {
  totalDownloads: number;
  todayDownloads: number;
  platformBreakdown: PublicStatsPlatformBreakdown;
  successRate: number;
}

export interface AdminCredentials {
  password: string;
}

export interface AdminToken {
  token: string;
  expiresAt: string;
}

export interface DailyCount {
  date: string;
  count: number;
}

export interface ApiHealth {
  name: string;
  platform: string;
  successRate: number;
  totalCalls: number;
  failures: number;
  enabled: boolean;
}

export type AdminStatsPlatformBreakdown = {
  tiktok?: number;
  youtube?: number;
  pinterest?: number;
};

export interface AdminStats {
  totalDownloads: number;
  todayDownloads: number;
  weekDownloads: number;
  platformBreakdown: AdminStatsPlatformBreakdown;
  successRate: number;
  activeJobs: number;
  pendingJobs: number;
  serverUptime: number;
  apiHealth: ApiHealth[];
  dailyCounts: DailyCount[];
  topErrors: string[];
}

export interface ErrorLog {
  id: string;
  timestamp: string;
  platform: string;
  url: string;
  error: string;
  apiUsed: string;
  retryCount: number;
}

export interface PlatformConfig {
  enabled: boolean;
  apiPriority: string[];
}

export interface AppConfig {
  tiktok: PlatformConfig;
  youtube: PlatformConfig;
  pinterest: PlatformConfig;
  rateLimitPerMinute: number;
  cacheExpiryMinutes: number;
}

export interface AppConfigUpdate {
  tiktok?: PlatformConfig;
  youtube?: PlatformConfig;
  pinterest?: PlatformConfig;
  rateLimitPerMinute?: number;
  cacheExpiryMinutes?: number;
}

export type DetectPlatformParams = {
url: string;
};

export type GetHistoryParams = {
limit?: number;
};

export type GetAdminLogsParams = {
limit?: number;
};