# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NovaSave API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminToken"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin statistics
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Admin statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/logs:
    get:
      operationId: getAdminLogs
      tags: [admin]
      summary: Get error logs
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Error logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorLog"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/jobs:
    get:
      operationId: getAdminJobs
      tags: [admin]
      summary: Get active jobs
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DownloadJob"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/config:
    get:
      operationId: getAdminConfig
      tags: [admin]
      summary: Get platform/API configuration
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppConfig"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateAdminConfig
      tags: [admin]
      summary: Update platform/API configuration
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppConfigUpdate"
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppConfig"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    MediaItem:
      type: object
      required: [url, quality, format, label]
      properties:
        url:
          type: string
        quality:
          type: string
        format:
          type: string
        label:
          type: string
        fileSize:
          type: ["integer", "null"]
    DownloadJob:
      type: object
      required: [jobId, status, url, platform, contentType, createdAt]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [pending, processing, ready, failed]
        url:
          type: string
        platform:
          type: string
        contentType:
          type: string
        title:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        author:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        mediaItems:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        error:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
        retryCount:
          type: integer
        apiUsed:
          type: ["string", "null"]
...
    AdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminToken:
      type: object
      required: [token, expiresAt]
      properties:
        token:
          type: string
        expiresAt:
          type: string
    DailyCount:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer
    ApiHealth:
      type: object
      required: [name, platform, successRate, totalCalls, failures, enabled]
      properties:
        name:
          type: string
        platform:
          type: string
        successRate:
          type: number
        totalCalls:
          type: integer
        failures:
          type: integer
        enabled:
          type: boolean
    AdminStats:
      type: object
      required: [totalDownloads, todayDownloads, weekDownloads, platformBreakdown, successRate, activeJobs, pendingJobs, serverUptime, apiHealth, dailyCounts, topErrors]
      properties:
        totalDownloads:
          type: integer
        todayDownloads:
          type: integer
        weekDownloads:
          type: integer
        platformBreakdown:
          type: object
          properties:
            tiktok:
              type: integer
            youtube:
              type: integer
            pinterest:
              type: integer
        successRate:
          type: number
        activeJobs:
          type: integer
        pendingJobs:
          type: integer
        serverUptime:
          type: number
        apiHealth:
          type: array
          items:
            $ref: "#/components/schemas/ApiHealth"
        dailyCounts:
          type: array
          items:
            $ref: "#/components/schemas/DailyCount"
        topErrors:
          type: array
          items:
            type: string
    ErrorLog:
      type: object
      required: [id, timestamp, platform, url, error, apiUsed, retryCount]
      properties:
        id:
          type: string
        timestamp:
          type: string
        platform:
          type: string
        url:
          type: string
        error:
          type: string
        apiUsed:
          type: string
        retryCount:
          type: integer
    PlatformConfig:
      type: object
      required: [enabled, apiPriority]
      properties:
        enabled:
          type: boolean
        apiPriority:
          type: array
          items:
            type: string
    AppConfig:
      type: object
      required: [tiktok, youtube, pinterest, rateLimitPerMinute, cacheExpiryMinutes]
      properties:
        tiktok:
          $ref: "#/components/schemas/PlatformConfig"
        youtube:
          $ref: "#/components/schemas/PlatformConfig"
        pinterest:
          $ref: "#/components/schemas/PlatformConfig"
        rateLimitPerMinute:
          type: integer
        cacheExpiryMinutes:
          type: integer
    AppConfigUpdate:
      type: object
      properties:
        tiktok:
          $ref: "#/components/schemas/PlatformConfig"
        youtube:
          $ref: "#/components/schemas/PlatformConfig"
        pinterest:
          $ref: "#/components/schemas/PlatformConfig"
        rateLimitPerMinute:
          type: integer
        cacheExpiryMinutes:
          type: integer