# lib/api-zod/src/generated/types/appConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NovaSave API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformConfig } from './platformConfig';

export interface AppConfig {
  tiktok: PlatformConfig;
  youtube: PlatformConfig;
  pinterest: PlatformConfig;
  rateLimitPerMinute: number;
  cacheExpiryMinutes: number;
}